04_djangorestframework/mysite/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
drf-writable-nested==0.6.3
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
